+++ unpublished/project_kit/lib/src/utils/token_storage/model/auth_token.dart
@@ -8,5 +8,8 @@ class AuthToken extends OAuth2Token {
     super.expiresIn,
     super.refreshToken,
     super.scope,
+    this.userId,
   });
+
+  final int? userId;
 }