+++ unpublished/project_kit/lib/src/utils/token_storage/security_token_storage.dart
@@ -14,8 +14,11 @@ part 'security_token_storage.g.dart';
 class UserChanged extends _$UserChanged {
   @override
   void build() {
-    ref.listen(securityTokenStorageProvider, (_, token) {
-      if (token.requireValue != null) ref.invalidateSelf();
+    ref.listen(securityTokenStorageProvider,
+        (asyncPrevToken, asyncCurrentToken) {
+      final prevUserId = asyncPrevToken?.requireValue?.userId;
+      final currentUserId = asyncCurrentToken.requireValue?.userId;
+      if (prevUserId != currentUserId) ref.invalidateSelf();
     });
   }
 
@@ -33,17 +36,26 @@ bool userAuthorized(UserAuthorizedRef ref) {
 @Riverpod(keepAlive: true)
 class SecurityTokenStorage extends _$SecurityTokenStorage
     implements IRef, TokenStorage<AuthToken> {
+  @Deprecated('Use _encryptedStorage instead')
   static const _storage = FlutterSecureStorage();
+  static const _encryptedStorage = FlutterSecureStorage(
+    aOptions: AndroidOptions(
+      encryptedSharedPreferences: true,
+    ),
+  );
   static const _refreshKey = 'refreshToken';
   static const _accessKey = 'accessToken';
 
   @override
   Future<AuthToken?> build() async {
-    await read();
+    final token = await read();
+    return token;
   }
 
   @override
   Future<void> delete() async {
+    await _encryptedStorage.delete(key: _refreshKey);
+    await _encryptedStorage.delete(key: _accessKey);
     await _storage.delete(key: _refreshKey);
     await _storage.delete(key: _accessKey);
     setData(null);
@@ -52,10 +64,18 @@ class SecurityTokenStorage extends _$SecurityTokenStorage
   @override
   Future<AuthToken?> read() async {
     try {
-      final refreshToken = await _storage.read(key: _refreshKey);
-      final accessToken = await _storage.read(key: _accessKey);
+      var refreshToken = await _storage.read(key: _refreshKey);
+      var accessToken = await _storage.read(key: _accessKey);
 
-      if (refreshToken == null || accessToken == null) return null;
+      if (refreshToken == null || accessToken == null) {
+        refreshToken = await _encryptedStorage.read(key: _refreshKey);
+        accessToken = await _encryptedStorage.read(key: _accessKey);
+        if (refreshToken == null || accessToken == null) return null;
+        return null;
+      }
+
+      await _encryptedStorage.write(key: _refreshKey, value: refreshToken);
+      await _encryptedStorage.write(key: _accessKey, value: accessToken);
 
       return AuthToken(
         accessToken: accessToken,
@@ -63,14 +83,14 @@ class SecurityTokenStorage extends _$SecurityTokenStorage
       );
     } catch (e, s) {
       logger.error(title: 'TokenStorage', error: e, stack: s);
-      return null;
+      rethrow;
     }
   }
 
   @override
   Future<void> write(AuthToken token) async {
-    await _storage.write(key: _refreshKey, value: token.refreshToken);
-    await _storage.write(key: _accessKey, value: token.accessToken);
+    await _encryptedStorage.write(key: _refreshKey, value: token.refreshToken);
+    await _encryptedStorage.write(key: _accessKey, value: token.accessToken);
 
     setData(
       AuthToken(